JAMES-3492, https://github.com/apache/james-project/pull/288#discussion_r571079820
